Move URL handler code into its own file
